accept hQmgi options

+++ includes/ft_ls.h
@@ -6,7 +6,7 @@
 /*   By: slindgre <slindgre@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/01 20:57:11 by gloras-t          #+#    #+#             */
-/*   Updated: 2020/12/25 22:32:59 by slindgre         ###   ########.fr       */
+/*   Updated: 2020/12/26 14:32:19 by slindgre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,8 +20,8 @@ int		ft_printf(const char *restrict format, ...);
 /*
 ** parse_args.c
 */
-int		parse_options(char *arg, uint8_t *options);
-void	parse_args(int argc, char *argv[], t_list **paths, uint8_t *options);
+int		parse_options(char *arg, uint16_t *options);
+void	parse_args(int argc, char *argv[], t_list **paths, uint16_t *options);
 
 /*
 ** comparsion_1.c
@@ -41,7 +41,7 @@ int		compare_by_size_desc(t_list *a, t_list *b);
 ** path_utils.c
 */
 int		verify_paths(t_list *paths, t_list **dirs, t_list **files,
-		uint8_t options);
+		uint16_t options);
 void	ft_lst_free_file(void *elem, size_t content_size);
 int		print_error(int err, char *path, int mode);
 void	map_to_file(t_stat buf, char *path, char *prefix, t_file *file);
@@ -50,19 +50,19 @@ int		is_dot_path(char *path);
 /*
 ** sort_list.c
 */
-void	sort_list(t_list **list, uint8_t options);
+void	sort_list(t_list **list, uint16_t options);
 
 /*
 ** print_file.c
 */
-void	print_file_info(t_file *file, uint8_t options);
-int		print_files(t_list *files, uint8_t options, int mode);
+void	print_file_info(t_file *file, uint16_t options);
+int		print_files(t_list *files, uint16_t	options, int mode);
 
 /*
 ** print_dir.c
 */
 t_list	*read_dir(t_file *dir);
-void	print_dirs(t_list *dirs, uint8_t options, int step);
+void	print_dirs(t_list *dirs, uint16_t options, int step);
 
 /*
 ** utils.c
